chore: 1. 스타일과 로직 분리 2. index.ts 역할 재설정 3. 루트 경로 순서 재설정
