+++ client/src/pages/main-page/components/Footer.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import Wrapper from './FooterStyle';
+
+const Footer: React.FC = () => {
+  return <Wrapper>푸터입니다.</Wrapper>;
+};
+
+export default Footer;
