+++ client/src/pages/main-page/components/FooterStyle.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const Wrapper = styled.section`
+  position: absolute;
+  top: 100%;
+  left: 50%;
+  transform: translate(-50%, -100%);
+  border: 1px solid;
+`;
+
+export default Wrapper;
