--- client/src/pages/main-page/components/Header/index.style.tsx
@@ -1,21 +1,0 @@
-import styled from 'styled-components';
-
-const LinkBtn = styled.button`
-  /* position: absolute; */
-  /* left: 40vw;
-  top: 50vh; */
-  margin-left: 30vw;
-`;
-
-const StyledHeader = styled.header`
-  display: flex;
-  height: 5vh;
-  width: 100vw;
-  /* top: 10vh;
-  left: 10vw; */
-  justify-items: center;
-  align-items: center;
-  position: absolute;
-`;
-
-export { LinkBtn, StyledHeader };
