+++ client/src/pages/main-page/components/Main.tsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import Wrapper from './MainStyle';
+import riceCakeImg from '../../../assets/images/riceCake.png';
+
+const Main: React.FC = () => {
+  return (
+    <Wrapper>
+      <div>
+        <img src={riceCakeImg} />
+      </div>
+      <div>
+        <img src={riceCakeImg} />
+      </div>
+      <div>
+        <img src={riceCakeImg} />
+      </div>
+      <div>
+        <img src={riceCakeImg} />
+      </div>
+      <div>
+        <img src={riceCakeImg} />
+      </div>
+      <div>
+        <img src={riceCakeImg} />
+      </div>
+    </Wrapper>
+  );
+};
+
+export default Main;
