+++ client/src/pages/main-page/components/index.tsx
@@ -1,0 +1,5 @@
+import Header from './Header';
+import Main from './Main';
+import Footer from './Footer';
+
+export { Header, Main, Footer };
